done films list controller
